improve UI and componentization
