+++ apps/frontend/src/components/beach-card/beach-info-section/style.module.scss
@@ -1,0 +1,38 @@
+@import 'src/styles/imports';
+
+
+.title {
+  font-size: 1.5rem;
+  font-weight: 600;
+  margin-left: .5rem;
+  text-align: right;
+}
+
+.beachcard-content {
+  padding: 0.75rem;
+  display: block;
+  position: absolute;
+  background-color: rgba(255, 255, 255, 1);
+  border-top-right-radius: 4px;
+  border-bottom-right-radius: 4px;
+  border-radius: 4px;
+  right: 15px;
+  bottom: 15px;
+  box-shadow: 0 1px 1px rgba(0,0,0,0.11),
+          0 2px 2px rgba(0,0,0,0.11),
+          0 4px 4px rgba(0,0,0,0.11),
+          0 8px 8px rgba(0,0,0,0.11),
+          0 16px 16px rgba(0,0,0,0.11),
+          0 32px 32px rgba(0,0,0,0.11);
+}
+
+.row {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+}
+
+.collection-date {
+  font-size: 0.65rem;
+  text-align: right;
+}
