+++ apps/frontend/src/components/beach-card/beach-status/style.module.scss
@@ -2,6 +2,7 @@
   background-color: grey;
   height: 1rem;
   width: 1rem;
+  display: block;
   border-radius: 100%;
 }
