+++ apps/frontend/src/components/beach-card/ecoli-status-label/style.module.scss
@@ -1,0 +1,10 @@
+
+.component {
+  font-size: 1rem;
+  margin-left: 5px;
+  margin-bottom: 4px;
+  margin-top: 4px;
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+}
