+++ apps/frontend/src/components/beach-card/index.tsx
@@ -1,52 +1,23 @@
 // import { useRouter } from 'next/router';
 import { Beach } from 'src/types/beaches';
-import { beachPositions } from '../../utils/beachPositions';
-import dayjs from 'dayjs';
-import relativeTime from 'dayjs/plugin/relativeTime';
 import styles from './style.module.scss';
-import BeachStatus from './beach-status';
 import Particle from './particle';
-// import { useRouter } from 'next/router';
-// import { beachIDToRouteName } from '../../utils/beachRouteMatch';
-
-dayjs.extend(relativeTime);
+import BeachInfoSection from './beach-info-section';
 
 interface Props {
   beach: Beach;
 }
 
 const BeachCard = ({ beach }: Props) => {
-  const { eColi,
-    collectionDate,
-    beachId } = beach;
-  const beachDisplayName = beachPositions(beachId).displayName;
+  const { eColi } = beach;
   // const router = useRouter();
 
-  const clickBeach = () => {
-    // const beachRouteName = beachIDToRouteName(beachId);
-    // router.push(`/beach/${beachRouteName}`, `/beach/${beachRouteName}`);
-  };
-
   return (
-    <div className={styles['beachcard']} onClick={clickBeach}>
+    <div className={styles['beachcard']}>
       <div className={styles['image-wrapper']}>
         <Particle ecoli={eColi} />
       </div>
-      <div className={styles['beachcard-content']}>
-        <div className={styles.title}>
-          {beachDisplayName}
-        </div>
-        <section className={styles.row}>
-          <BeachStatus eColi={eColi} />
-          <div className={styles.ecoli}>
-            {eColi && <> {eColi} E. coli </>}
-            {!eColi && <> Not Tested </>}
-          </div>
-        </section>
-        <div className={styles['collection-date']}>
-          {dayjs(collectionDate).fromNow()}
-        </div>
-      </div>
+      <BeachInfoSection beach={beach} />
     </div>
   );
 };
