+++ apps/frontend/src/components/home-page-header/style.module.scss
@@ -6,5 +6,8 @@
 }
 
 .title {
-  font-size: 1.5rem;
+  font-size: 2rem;
+  font-family: 'Jost', sans-serif;
+  margin-bottom: 0;
+  font-weight: 500;
 }
