+++ apps/frontend/src/pages/_document.tsx
@@ -7,6 +7,9 @@ export default function Document() {
       <Head />
       <body>
         <Main />
+        <link rel="preconnect" href="https://fonts.googleapis.com" />
+        <link rel="preconnect" href="https://fonts.gstatic.com" />
+        <link href="https://fonts.googleapis.com/css2?family=Jost:wght@200;400;500;700&display=swap" rel="stylesheet"></link>
         <NextScript />
       </body>
     </Html>
